+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Coming soon
+- A simpler configuration and startup helper with `NewRelicExporters.java`
 - Support for version 0.7.0 of OpenTelemetry Java
 
 ## [0.6.1] - 2020-07-xx